Some tweaks to go script architecture/structure
